Separating out the config entries
