+++ src/util.js
@@ -1,40 +1,6 @@
 const bitcoin = require("bitcoinjs-lib")
 const BigNumber = require('bignumber.js');
-
-
-export const NETWORKS = {
-  1: {
-    name: "mainnet",
-    etherscan: 'https://etherscan.io',
-    infura: 'http://mainnet.infura.io',
-    marketplace:'',
-    protobufjs:'http://protobufjs.singularitynet.io/'
-  },
-  3: {
-    name: "Ropsten",
-    etherscan: 'https://ropsten.etherscan.io',
-    infura: 'https://ropsten.infura.io',
-    marketplace:'',
-    protobufjs:'http://protobufjs.singularitynet.io/'
-  },
-  4: {
-    name: "Rinkeby",
-    etherscan: 'https://rinkeby.etherscan.io',
-    infura: 'https://rinkeby.infura.io',
-    marketplace:'',
-    protobufjs:'http://protobufjs.singularitynet.io/'
-  },
-  42: {
-    name: "Kovan",
-    etherscan: 'https://kovan.etherscan.io',
-    infura: 'https:/kovan.infura.io',
-    marketplace:'https://260r82zgt7.execute-api.us-east-1.amazonaws.com/kovan/',
-    protobufjs:'http://protobufjs.singularitynet.io/',
-    default:true
-  },
-};
-
-export const SERVICE_SPEC_PROVIDER_URL = "http://protobufjs.singularitynet.io"
+const { NETWORKS } = require("./networks.js") 
 
 export const AGENT_STATE = {
   "ENABLED": 0,
